+++ packages/website/components/LiveEditor/ComponentSource.tsx
@@ -13,9 +13,7 @@ import * as f36Components from '@contentful/f36-components';
 import { Multiselect } from '@contentful/f36-multiselect';
 import { NavList } from '@contentful/f36-navlist';
 import * as f36utils from '@contentful/f36-utils';
-import { Avatar, AvatarGroup } from '@contentful/f36-avatar';
 import { Header } from '@contentful/f36-header';
-import { Image } from '@contentful/f36-image';
 import { Layout } from '@contentful/f36-layout';
 import { useForm, useController } from 'react-hook-form';
 import { MdAccessAlarm } from 'react-icons/md';
@@ -41,10 +39,7 @@ const liveProviderScope = {
   ...f36Components,
   ...f36icons,
   ...f36utils,
-  Avatar, // Remove when avatar is added to f36-components
-  AvatarGroup, // Remove when avatar is added to f36-components
   Header, // Remove when added to f36-components
-  Image, // Remove when added to f36-components
   Layout, // Remove when added to f36-components
   Multiselect, // Remove when added to f36-components
   NavList, // Remove when added to f36-components