+++ CHANGELOG.md
@@ -8,6 +8,20 @@ The Changelog gives an overview of the changes we've made to Forma 36
 
 <!-- CHANGELOG:INSERT -->
 
+## 24-05-2024
+
+`@contentful/f36-components@4.65.7`
+
+- feat(changesets): move avatar from alpha package to automatic release [AHOY-3166]
+
+**F36 Avatar** `v4.65.7`
+
+- feat(changesets): move avatar from alpha package to automatic release [AHOY-3166]
+
+**F36 Image** `v4.65.7`
+
+- feat(changesets): move avatar from alpha package to automatic release [AHOY-3166]
+
 ## 14-05-2024
 
 **F36 Navbar** `v4.2.0`